+++ tempodb/tempodb_search_test.go
@@ -154,7 +154,7 @@ func advancedTraceQLRunner(t *testing.T, wantTr *tempopb.Trace, wantMeta *tempop
 		fmt.Sprintf("rootName=`%s`", wantMeta.RootTraceName),
 	}
 	totalSpans := 0
-	for _, b := range wantTr.Batches {
+	for _, b := range wantTr.ResourceSpans {
 		trueResourceC, falseResourceC := conditionsForAttributes(b.Resource.Attributes, "resource")
 		falseConditions = append(falseConditions, falseResourceC...)
 
@@ -1868,7 +1868,7 @@ func makeExpectedTrace() (
 	end = 1001
 
 	tr = &tempopb.Trace{
-		Batches: []*v1.ResourceSpans{
+		ResourceSpans: []*v1.ResourceSpans{
 			{
 				Resource: &v1_resource.Resource{
 					Attributes: []*v1_common.KeyValue{
@@ -2192,7 +2192,7 @@ func TestWALBlockGetMetrics(t *testing.T) {
 
 	// Write to wal
 	err = head.AppendTrace(common.ID{0x01}, &tempopb.Trace{
-		Batches: []*v1.ResourceSpans{
+		ResourceSpans: []*v1.ResourceSpans{
 			{
 				ScopeSpans: []*v1.ScopeSpans{
 					{